requirements.txt
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
psycopg2-binary>=2.9
sqlalchemy>=1.4
joblib>=1.2
lightgbm>=3.3.5
python-dotenv
fastapi
uvicorn
pytest